requirements.txt
docutils
pygments
pypiwin32
kivy_deps.sdl2
kivy_deps.glew
kivy_deps.gstreamer
kivy_deps.angle
Kivy==2.0.0rc3
arabic_reshaper
python-bidi
babel